src/main/java/sky/pro/java/course2/examinerservice/service/ExaminerServiceImpl.java
package sky.pro.java.course2.examinerservice.service;

import org.springframework.stereotype.Service;
import sky.pro.java.course2.examinerservice.domain.Question;
import sky.pro.java.course2.examinerservice.exceptions.QuestionInvalidAmountException;

import java.util.*;

@Service
public class ExaminerServiceImpl implements ExaminerService {
    QuestionService questionService;

    public ExaminerServiceImpl(QuestionService questionService) {
        this.questionService = questionService;
    }

    @Override
    public Collection<Question> getQuestions(int amount) {
        List<Question> questions = new ArrayList<>();
        if (amount < 1 || amount > 10) {
            throw new QuestionInvalidAmountException();
        }
        int counter = 0;
        while (counter != amount) {
            Question question = questionService.getRandomQuestion();
            if (!questions.contains(question)) {
                questions.add(question);
                counter ++;
            }
        }
        return questions;
    }
}
